refactor(a2a): reduce interop suite duplication
