+++ integrations/agntcy-a2a/README.md
@@ -155,6 +155,8 @@ task test:python-dotnet
 
 `task test` now runs the full `task test:rust-go` transport matrix plus `task test:rust-dotnet`, `task test:go-dotnet`, `task test:python-go`, `task test:rust-python`, and `task test:python-dotnet`.
 
+To keep the Taskfile compact, `task -l` now shows wildcard-backed task families such as `test:rust-go:*`, `test:rust-go:*:*`, and `test:rust-go:behavior:*` rather than one line per expanded matrix case. The concrete commands below still work unchanged.
+
 The user-facing task triggers are organized by scope:
 
 - Full suites:
@@ -357,8 +359,8 @@ That one behavior entry is expanded into multiple specs because the suite wrappe
 To add a new shared behavior:
 
 1. Add a new method to `interopHarness` in `tests/interop_behaviors_test.go`.
-2. Implement that method in each harness that should participate, such as `goSDKHarness`, `rustProbeHarness`, `dotNetProbeHarness`, and `pythonProbeHarness`.
-3. Add a new entry to `sharedInteropBehaviorSpecs` with a stable label. If the behavior should be runnable as a first-class filtered target like the current behavior slices, add matching Taskfile targets and document them here.
+2. Implement that method in each harness that should participate. The native Go path lives in `goSDKHarness` in `tests/native_go_harness_test.go`; the external SDK probe paths are wired through `newRustProbeHarness(...)`, `newDotNetProbeHarness(...)`, and `newPythonProbeHarness(...)` in `tests/interop_behaviors_test.go`.
+3. Add a new entry to `sharedInteropBehaviorSpecs` with a stable label. If the behavior should be runnable as a first-class filtered target like the current behavior slices, extend the wildcard-backed behavior task families in `Taskfile.yml` and document the new label here.
 4. If the Rust, .NET, or Python harnesses need their own focused scenario selection, add a new `probeScenario` in `tests/interop_shared_test.go`, pass it through `tests/interop_launchers_test.go`, and implement the scenario in the matching probe binary:
    `fixtures/rust/src/bin/interop-rust-probe.rs`
    `fixtures/dotnet/InteropProbe/Program.cs`
